Refactor main __init__.py of server and store to backend common
