+++ vantage6-algorithm-store/vantage6/algorithm/store/__init__.py
@@ -9,6 +9,7 @@ store to a vantage6 server.
 
 import os
 from gevent import monkey
+from vantage6.backend.common.permission import RuleNeed
 
 # This is a workaround for readthedocs
 if not os.environ.get("READTHEDOCS"):
@@ -24,14 +25,23 @@ import datetime
 
 from http import HTTPStatus
 from werkzeug.exceptions import HTTPException
-from flask import Flask, make_response, request, send_from_directory, Request, Response
+from flask import (
+    Flask,
+    make_response,
+    request,
+    send_from_directory,
+    Request,
+    Response,
+    current_app,
+)
 from flask_cors import CORS
 from flask_marshmallow import Marshmallow
 from flask_restful import Api
-from flask_principal import Principal
-from flasgger import Swagger
+from flask_principal import Principal, Identity, identity_changed
 from pathlib import Path
 from keycloak import KeycloakOpenID
+from flask_jwt_extended import JWTManager
+
 
 from vantage6.common import logger_name
 from vantage6.common.globals import APPNAME, DEFAULT_API_PATH
@@ -80,68 +90,16 @@ class AlgorithmStoreApp(Vantage6App):
     def __init__(self, ctx: AlgorithmStoreContext) -> None:
         """Create a vantage6 algorithm store application."""
 
-        self.ctx = ctx
-
-        # validate that the required environment variables are set
-        self.validate_required_env_vars()
-
-        # initialize, configure Flask
-        self.app = Flask(
-            SERVER_MODULE_NAME,
-            root_path=Path(__file__),
-            template_folder=Path(__file__).parent / "templates",
-            static_folder=Path(__file__).parent / "static",
-        )
-        self.debug: dict = self.ctx.config.get("debug", {})
-        self.configure_flask()
-
-        # Setup SQLAlchemy and Marshmallow for marshalling/serializing
-        self.ma = Marshmallow(self.app)
-
-        # Setup Principal, granular API access manegement
-        self.principal = Principal(self.app, use_sessions=False)
-
-        # Enable cross-origin resource sharing
-        cors_allowed_origins = self.ctx.config.get("cors_allowed_origins", "*")
-        self.cors = CORS(
-            self.app,
-            resources={r"/*": {"origins": cors_allowed_origins}},
-        )
-
-        # SWAGGER documentation
-        self.swagger = Swagger(self.app, template={})
-
-        # setup Flask mail client
-        self.mail = MailService(self.app)
+        super().__init__(ctx, SERVER_MODULE_NAME)
 
         # setup the permission manager for the API endpoints
         self.permissions = PermissionManager(RESOURCES_PATH, RESOURCES, DefaultRole)
 
-        # sync policies with the database
-        self.setup_policies(self.ctx.config)
-
-        # Api - REST JSON-rpc
-        self.api = Api(self.app)
-        self.configure_api()
+        # Load API resources
         self.load_resources()
 
-        # set the server version
-        self.__version__ = __version__
-
-        # set environment variable for dev environment
-        host_uri = self.ctx.config.get("dev", {}).get("host_uri")
-        if host_uri:
-            os.environ[HOST_URI_ENV] = host_uri
-
-        # TODO v5+ remove this - for backwards compatibility of v4.6 with v4.3-4.5 we
-        # are setting algorithms with empty values for the `submitted` field (new in
-        # v4.6) to approved
-        for algorithm in db.Algorithm.get():
-            if not algorithm.submitted_at:
-                algorithm.status = ReviewStatus.APPROVED.value
-                algorithm.submitted_at = datetime.datetime.now(datetime.timezone.utc)
-                algorithm.approved_at = datetime.datetime.now(datetime.timezone.utc)
-                algorithm.save()
+        # sync policies with the database
+        self.setup_policies(self.ctx.config)
 
         if self.ctx.config.get("dev", {}).get("disable_review", False):
             self.setup_disable_review()
@@ -150,144 +108,59 @@ class AlgorithmStoreApp(Vantage6App):
 
     def configure_flask(self) -> None:
         """Configure the Flask settings of the vantage6 algorithm store."""
-        # TODO there is some duplicate code with the server here, check if it can be
-        # refactored
-
-        # let us handle exceptions
-        self.app.config["PROPAGATE_EXCEPTIONS"] = True
-
-        # Open Api Specification (f.k.a. swagger)
-        self.app.config["SWAGGER"] = {
-            "title": f"{APPNAME} algorithm store",
-            "uiversion": "3",
-            "openapi": "3.0.0",
-            "version": __version__,
-        }
-
-        # Mail settings
-        mail_config = self.ctx.config.get("smtp", {})
-        self.app.config["MAIL_PORT"] = mail_config.get("port", 1025)
-        self.app.config["MAIL_SERVER"] = mail_config.get("server", "localhost")
-        self.app.config["MAIL_USERNAME"] = mail_config.get(
-            "username", DEFAULT_SUPPORT_EMAIL_ADDRESS
-        )
-        self.app.config["MAIL_PASSWORD"] = mail_config.get("password", "")
-        self.app.config["MAIL_USE_TLS"] = mail_config.get("MAIL_USE_TLS", True)
-        self.app.config["MAIL_USE_SSL"] = mail_config.get("MAIL_USE_SSL", False)
-
-        debug_mode = self.debug.get("flask", False)
-        if debug_mode:
-            log.debug("Flask debug mode enabled")
-        self.app.debug = debug_mode
-
-        def _get_request_path(request: Request) -> str:
-            """
-            Return request extension of request URL, e.g.
-            http://localhost:7601/api/task/1 -> api/task/1
-
-            Parameters
-            ----------
-            request: Request
-                Flask request object
-
-            Returns
-            -------
-            string:
-                The endpoint path of the request
-            """
-            return request.url.replace(request.url_root, "")
-
-        # before request
-        @self.app.before_request
-        def do_before_request():
-            """Before every flask request method."""
-            # Add log message before each request
-            log.debug(
-                "Received request: %s %s", request.method, _get_request_path(request)
-            )
-
-            # This will obtain a (scoped) db session from the session factory
-            # that is linked to the flask request global `g`. In every endpoint
-            # we then can access the database by using this session. We ensure
-            # that the session is removed (and uncommited changes are rolled
-            # back) at the end of every request.
-            DatabaseSessionManager.new_session()
-
-        @self.app.after_request
-        def remove_db_session(response):
-            """After every flask request.
-
-            This will close the database session created by the
-            `before_request`.
-            """
-            DatabaseSessionManager.clear_session()
-            return response
-
-        @self.app.errorhandler(HTTPException)
-        def error_remove_db_session(error: HTTPException):
-            """In case an HTTP-exception occurs during the request.
-
-            It is important to close the db session to avoid having dangling
-            sessions.
-            """
-            if error.code == 404:
-                log.debug("404 error for route '%s'", _get_request_path(request))
-            else:
-                log.warning("HTTP Exception occured during request")
-                log.warning("%s", traceback.format_exc())
-            DatabaseSessionManager.clear_session()
-            return error.get_response()
-
-        @self.app.errorhandler(Exception)
-        def error2_remove_db_session(error):
-            """In case an exception occurs during the request.
-
-            It is important to close the db session to avoid having dangling
-            sessions.
-            """
-            log.exception("Exception occured during request")
-            DatabaseSessionManager.clear_session()
-            return {
-                "msg": "An unexpected error occurred on the server!"
-            }, HTTPStatus.INTERNAL_SERVER_ERROR
-
-        @self.app.route("/robots.txt")
-        def static_from_root():
-            return send_from_directory(self.app.static_folder, request.path[1:])
+        self._configure_flask_base(DatabaseSessionManager)
 
     def configure_api(self) -> None:
         """Define global API output and its structure."""
+        self._configure_api_base(BaseHATEOASModelSchema, Base)
 
-        # helper to create HATEOAS schemas
-        BaseHATEOASModelSchema.api = self.api
+    def configure_jwt(self):
+        """Configure JWT authentication."""
 
-        # whatever you get try to json it
-        @self.api.representation("application/json")
-        # pylint: disable=unused-argument
-        def output_json(
-            data: Base | list[Base], code: HTTPStatus, headers: dict = None
-        ) -> Response:
+        @self.jwt.user_lookup_loader
+        def user_lookup_loader(jwt_payload: dict, jwt_headers: dict) -> db.User:
             """
-            Return jsonified data for request responses.
+            Load the user, node or container instance from the JWT payload.
 
             Parameters
             ----------
-            data: Base | list[Base]
-                The data to be jsonified
-            code: HTTPStatus
-                The HTTP status code of the response
-            headers: dict
-                Additional headers to be added to the response
+            jwt_payload: dict
+                The JWT payload
+            jwt_headers: dict
+                The JWT headers
+
+            Returns
+            -------
+            db.User:
+                The user identity.
             """
+            identity = jwt_headers["sub"]
+            auth_identity = Identity(identity)
 
-            if isinstance(data, Base):
-                data = jsonable(data)
-            elif isinstance(data, list) and len(data) and isinstance(data[0], Base):
-                data = jsonable(data)
+            user = db.User.get_by_keycloak_id(identity)
+            if not user:
+                raise Exception("No user found for keycloak id %s", identity)
 
-            resp = make_response(json.dumps(data), code)
-            resp.headers.extend(headers or {})
-            return resp
+            # add role permissions
+            for role in user.roles:
+                for rule in role.rules:
+                    auth_identity.provides.add(
+                        RuleNeed(
+                            name=rule.name,
+                            operation=rule.operation,
+                        )
+                    )
+            for rule in user.rules:
+                auth_identity.provides.add(
+                    RuleNeed(
+                        name=rule.name,
+                        operation=rule.operation,
+                    )
+                )
+            identity_changed.send(
+                current_app._get_current_object(), identity=auth_identity
+            )
+            return user
 
     def load_resources(self) -> None:
         """Import the modules containing API resources."""
@@ -305,26 +178,6 @@ class AlgorithmStoreApp(Vantage6App):
         for res in RESOURCES:
             module = importlib.import_module("vantage6.algorithm.store.resource." + res)
             module.setup(self.api, api_path, services)
-
-    @staticmethod
-    def _add_default_roles() -> None:
-        for role in get_default_roles():
-            if not db.Role.get_by_name(role["name"]):
-                log.warning("Creating new default role %s", role["name"])
-                new_role = db.Role(
-                    name=role["name"],
-                    description=role["description"],
-                    rules=role["rules"],
-                    is_default_role=True,
-                )
-                new_role.save()
-            else:
-                current_role = db.Role.get_by_name(role["name"])
-                # check that the rules are the same. Use set() to compare without order
-                if set(current_role.rules) != set(role["rules"]):
-                    log.warning("Updating default role %s with new rules", role["name"])
-                    current_role.rules = role["rules"]
-                    current_role.save()
 
     def setup_policies(self, config: dict) -> None:
         """
@@ -418,7 +271,7 @@ class AlgorithmStoreApp(Vantage6App):
         Before server is really started, some database settings are checked and
         (re)set where appropriate.
         """
-        self._add_default_roles()
+        self._add_default_roles(get_default_roles(), db)
 
         # add whitelisted server and root user from config file if they do not exist
         if root_user := self.ctx.config.get("root_user", {}):
@@ -477,27 +330,6 @@ class AlgorithmStoreApp(Vantage6App):
                 " the database. This means no-one can alter resources on this server."
             )
         return self
-
-    # TODO v5+ refactor with duplicate code in server
-    def _add_keycloak_id_to_super_user(self, user: db.User) -> None:
-        """
-        Add a keycloak id to the super user.
-        """
-        keycloak_openid = KeycloakOpenID(
-            server_url=os.environ.get(RequiredServerEnvVars.KEYCLOAK_URL.value),
-            client_id=os.environ.get(RequiredServerEnvVars.KEYCLOAK_ADMIN_CLIENT.value),
-            realm_name=os.environ.get(RequiredServerEnvVars.KEYCLOAK_REALM.value),
-            client_secret_key=os.environ.get(
-                RequiredServerEnvVars.KEYCLOAK_ADMIN_CLIENT_SECRET.value
-            ),
-        )
-        token = keycloak_openid.token(
-            os.environ.get(RequiredServerEnvVars.KEYCLOAK_ADMIN_USERNAME.value),
-            os.environ.get(RequiredServerEnvVars.KEYCLOAK_ADMIN_PASSWORD.value),
-        )
-        decoded_token = keycloak_openid.decode_token(token["access_token"])
-        user.keycloak_id = decoded_token["sub"]
-        user.save()
 
 
 def run_server(config: str, system_folders: bool = True) -> AlgorithmStoreApp:
